# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Бабкосчёт API
  version: "1.0.0"
  description: |
    Личные финансы — REST API для счетов, операций, бюджетов, отчётов,
    регулярных платежей и целей накопления.

    **Аутентификация:** access JWT (15 мин) в `Authorization: Bearer <token>`
    + refresh-token в HttpOnly cookie (30 дней). Регистрируйтесь через
    `POST /auth/register`, потом передавайте access в заголовке.

    **Деньги** — всегда строки в JSON (`"1500.00"`), никогда не float.
    Дефолтная валюта `RUB`.

    **RLS:** все запросы исполняются под ролью `app_user` с
    `app.current_user_id = JWT.sub`. Чужие данные физически невидимы.
servers:
  - url: /api/v1
    description: Same-origin (Vite proxy в dev, nginx в prod)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /reports/monthly-summary:
    get:
      tags: [Reports]
      summary: Доход / расход / нетто помесячно
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to,   in: query, required: true, schema: { type: string, format: date } }
      responses: { "200": { description: OK } }
  /reports/spending-by-category:
    get:
      tags: [Reports]
      summary: Расходы сгруппированные по категории
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to,   in: query, required: true, schema: { type: string, format: date } }
      responses: { "200": { description: OK } }
  /reports/cash-flow:
    get:
      tags: [Reports]
      summary: Cash-flow по дням / неделям / месяцам
      parameters:
        - { name: from,        in: query, required: true,  schema: { type: string, format: date } }
        - { name: to,          in: query, required: true,  schema: { type: string, format: date } }
        - { name: granularity, in: query, required: false, schema: { type: string, enum: [day, week, month], default: month } }
      responses: { "200": { description: OK } }

  # ── Goals ──────────────────────────────────────────────────────────────
...